Basic Docker support for pull request stats
Now we have a basic Docker container running for the action, let's have
it run the ruby script.

This is out smallest next step, we don't plan on keeping the git
checkout inside the script as we think there is no need, but let's get
the basics down first.

We want to see if the action can see the docker file outside of it's
directory.