Add JS scripts for fetching historical data
It seems like the decision has been made to rewrite this all in Ruby,
but I thought it would be good to at least put these scripts in the
codebase (even if we just leave this as a draft PR) so we can copy them
across.

Sorry for nesting a whole project inside another project!