PullRequest does data transformation for Influx
Move the method into the PullRequest class, which now has all the data
required to combine and transform the PR data with the release data in
the shape we want to store in InfluxDB.

Remove the gigantic comment - it makes the code harder to read, and not
all comments are relevant to the code, because they refer to writing
Influx queries. The comments will resurface as wiki documentation.