WIP List repositories without teams
Repositories that don't give access to the tech team raise an error
when the dxw-rails-user tries to inspect their teams, so rather than
being able to list explicitly which repositories don't give access to
which teams, we have to rely on the proxy indicator of listing repos for
which the teams couldn't be accessed.

This branch should be its own repo, but it was easier to kickstart the
spike from an existing one.